requirements.txt
fastapi
python-multipart
jinja2
transformers
torch
scipy
python-dotenv
uvicorn
